requirements.txt
flask==2.3.2
Werkzeug<3
mysql-connector-python==8.3.0
python-dotenv==1.0.0

# Testing dependencies
pytest==7.4.0
pytest-flask==1.2.0
pytest-cov==4.1.0
pytest-mock==3.11.1